Tackle a hanging TODO, and do some cleanup
Created a parent class for Shipper and ShipTo called Client which wraps
some of the common behavior for those two entites.  Also standardized
the initialization of children of the Client class, having it take a
single hash rather than arguments followed by an options hash.  Added
simple presence validation to ensure that certain required values were
always passed in on init.  Renamed Ups::ShipTo, (which kind of sounds
like an address), to Ups::Consignee, which is both the correct name
_and_ the name used by UPS internally for the receiver of a package.

Updated the specs for Ups::Shipper, modernized them, slimmed them down,
and generally contributed to their hotness...

TODO: Next step is to apply a similar treatment to the Ups::Consignee
spec. Also need to clean up the Address class so that we can finish
writing specs for the Shipper and Consignee