Clean up consignee specs and add address features
Consignee Specs are a mirror of Shipper specs, but obviously focusing on
the Consignee's individual attributes. Added a feature to the Address
class wherein it is possible to give it a differently named root node,
(for that case where you need an address to be represented as a
ShipFrom). It may be possible that we want to do this for every to_xml?